basic venting functionality
Currently just vents whenever we're standing on top of a vent. Will pick a random vent and check if there are any players nearby. If there are no players it will assume it is safe to exit and leave the vent.

Also fixes a bug in Vent_Patches.cs since vent.Use() apparently never fires.